feat: got satisfied with the base of the game
Base of the game which i'm gonna build upon. At this State we have: a field, randomly spawning enemies which can be killed in a dynamic way. Some global UI which interacts with game's state (Menu button, score) and local UI (damage popups) which is also interactive inside the game.